+++ src/app/app.controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AppController } from './app.controller';
+import { AppController } from '../app.controller';
 import { AppService } from './app.service';
 
 describe('AppController', () => {
